Adding first pass TransportItem code - config parsing, metadata, variant handling, two default textures. Changing dimensions to register in the map as they are discovered and validated.

+++ src/main/java/nil/simpledim/config/DimensionInfo.java
@@ -12,7 +12,7 @@ public class DimensionInfo {
 	public String name;
 	public int dimensionId;
 	public DimensionType type;
-	public Integer seedOverride;
+	public Long seedOverride;
 	public ChunkCoordinates spawnPoint;
 	public String superflatGenerator;
 	public BiomeGenBase biome;
